tailscale_integrated/requirements.txt
flask
requests
qrcode
python-dotenv
